+++ app/builders/daggerheart_character/classes/bard_builder.rb
@@ -11,6 +11,12 @@ module DaggerheartCharacter
         result[:traits] = { 'str' => -1, 'agi' => 0, 'fin' => 1, 'ins' => 0, 'pre' => 2, 'know' => 1 }
         result
       end
+
+      def equip(character:)
+        Character::Item.create(character: character, item: Daggerheart::Item.find_by(slug: 'rapier'), ready_to_use: true)
+        Character::Item.create(character: character, item: Daggerheart::Item.find_by(slug: 'small_dagger'), ready_to_use: true)
+        Character::Item.create(character: character, item: Daggerheart::Item.find_by(slug: 'gambeson_armor'), ready_to_use: true)
+      end
     end
   end
 end