+++ app/builders/daggerheart_character/classes/guardian_builder.rb
@@ -11,6 +11,11 @@ module DaggerheartCharacter
         result[:traits] = { 'str' => 2, 'agi' => 1, 'fin' => -1, 'ins' => 0, 'pre' => 1, 'know' => 0 }
         result
       end
+
+      def equip(character:)
+        Character::Item.create(character: character, item: Daggerheart::Item.find_by(slug: 'battleaxe'), ready_to_use: true)
+        Character::Item.create(character: character, item: Daggerheart::Item.find_by(slug: 'chainmail_armor'), ready_to_use: true)
+      end
     end
   end
 end