+++ app/commands/homebrew_context/daggerheart/add_feat_command.rb
@@ -3,6 +3,7 @@
 module HomebrewContext
   module Daggerheart
     class AddFeatCommand < BaseCommand
+      # rubocop: disable Metrics/BlockLength
       use_contract do
         config.messages.namespace = :homebrew_feat
 
@@ -15,7 +16,7 @@ module HomebrewContext
           required(:title).filled(:string, max_size?: 50)
           required(:description).filled(:string, max_size?: 250)
           required(:origin).filled(Origins)
-          required(:origin_value).filled(:string, :uuid_v4?)
+          required(:origin_value).filled(:string)
           required(:kind).filled(Kinds)
           optional(:limit).filled(:integer, gteq?: 1)
           optional(:limit_refresh).filled(Limits)
@@ -25,14 +26,19 @@ module HomebrewContext
         rule(:origin, :origin_value, :user) do
           origin =
             case values[:origin]
-            when 'ancestry' then ::Daggerheart::Homebrew::Race.find_by(user_id: values[:user].id, id: values[:origin_value])
-            when 'class' then ::Daggerheart::Homebrew::Speciality.find_by(user_id: values[:user].id, id: values[:origin_value])
+            when 'ancestry'
+              ::Daggerheart::Homebrew::Race.find_by(user_id: values[:user].id, id: values[:origin_value]) ||
+                ::Daggerheart::Character.heritage_info(values[:origin_value])
+            when 'class'
+              ::Daggerheart::Homebrew::Speciality.find_by(user_id: values[:user].id, id: values[:origin_value]) ||
+                ::Daggerheart::Character.class_info(values[:origin_value])
             end
           next if origin
 
           key(:origin).failure(:not_found)
         end
       end
+      # rubocop: enable Metrics/BlockLength
 
       private