+++ app/services/characters_context/refresh_feats.rb
@@ -14,7 +14,7 @@ module CharactersContext
 
     def find_available_feats(character)
       all_feats = filter_available_feats(character)
-      exclude_feats = all_feats.pluck(:exclude).flatten.uniq
+      exclude_feats = all_feats.pluck(:exclude).flatten.uniq.compact
       all_feats.reject { |item| item.slug.in?(exclude_feats) }
     end