+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - grouping features by origin
 - adding homebrew feats for default races/classes
 
+### Fixed
+- null attack value
+
 ## [0.2.11] - 2025-07-29
 ### Added
 - starting equipment for Daggerheart